Upgrade recursion/... from u32 to u64
This avoids overflow in benchmark.rs.

+++ recursion/snip/fib_naive.rs
@@ -1,4 +1,4 @@
-fn fib_naive(n: u32) -> u32 {
+fn fib_naive(n: u64) -> u64 {
     if n < 2 {
         n
     } else {
