+++ recursion/snip/fibonacci_hof.rs
@@ -1,4 +1,6 @@
-fn fibonacci_hof(n: u32) -> u32 {
+use std::iter::successors;
+
+fn fibonacci_hof(n: u64) -> u64 {
     successors(Some((0, 1)), |&(a, b)| Some((b, a + b)))
         .nth(n as usize)
         .expect("Fibonacci sequence should be inexhaustible")
